refactor(cli): split SPEC-92 god file into per-command modules
