+++ src/tests/units/main/executeStatus.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { executeStatus, type StatusDeps } from '../../../main/cli.js';
+import { executeStatus, type StatusDeps } from '@/main/commands/status.command.js';
 import type { QueryStatusDependencies } from '@/modules/cli-configuration/usecases/cli/queryStatus.usecase.js';
 import { createPidFileContent } from '../../factories/pidFileContent.factory.js';
