+++ src/tests/units/main/executeStop.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { executeStop, type StopDeps } from '../../../main/cli.js';
+import { executeStop, type StopDeps } from '@/main/commands/stop.command.js';
 import type { StopDaemonDependencies } from '@/modules/cli-configuration/usecases/cli/stopDaemon.usecase.js';
 import { createPidFileContent } from '../../factories/pidFileContent.factory.js';
